refactor(app): extract content persistence from app model
Move quit-time hooks/tasks persistence out of app.go so the grandfathered file-length ceiling is no longer at risk. Behavior is unchanged; this is a mechanical extraction for #1145.

References #1240.
